import fix and small packages refactoring
